Use size_t in GCIdleTimeHandler to fix undefined behaviour.
BUG=
R=bmeurer@chromium.org

Review URL: https://codereview.chromium.org/490943002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23248 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap/heap.cc
@@ -4265,8 +4265,10 @@ void Heap::MakeHeapIterable() {
 
 
 void Heap::AdvanceIdleIncrementalMarking(int idle_time_in_ms) {
-  intptr_t step_size = GCIdleTimeHandler::EstimateMarkingStepSize(
-      idle_time_in_ms, tracer_.IncrementalMarkingSpeedInBytesPerMillisecond());
+  intptr_t step_size =
+      static_cast<size_t>(GCIdleTimeHandler::EstimateMarkingStepSize(
+          idle_time_in_ms,
+          tracer_.IncrementalMarkingSpeedInBytesPerMillisecond()));
 
   incremental_marking()->Step(step_size,
                               IncrementalMarking::NO_GC_VIA_STACK_GUARD, true);
